Change Total type to float64

+++ 4-order-api/pkg/di/models.go
@@ -10,7 +10,7 @@ type Order struct {
 	gorm.Model
 	UserID uint
 	User   users.User `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
-	Total  int        `json:"total"`
+	Total  float64    `json:"total"`
 	Items  []OrderItem
 }
 
@@ -21,5 +21,5 @@ type OrderItem struct {
 	ProductID uint             `json:"productId" gorm:"uniqueIndex:idx_order_item_order_id_product_id"` //gorm:"uniqueIndex:idx_order_item_order_id_product_id"
 	Product   products.Product `gorm:"PRELOAD:false,constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 	Count     int              `json:"count"`
-	Price     int              `json:"price"`
+	Price     float64          `json:"price"`
 }
